fix: more refactoring, sticky footer, select details from weights page
